SmartScheduler_v2.2/requirements.txt
asgiref==3.5.2
astroid==2.11.7
azure-core==1.24.2
azure-identity==1.10.0
bandit==1.7.4
boto3==1.26.11
botocore==1.29.11
cachetools==5.2.0
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloc==0.2.5
colorama==0.4.3
colored==1.4.3
commonmark==0.9.1
cryptography==37.0.2
Deprecated==1.2.13
dill==0.3.5.1
dj-database-url==1.0.0
Django==3.2.16
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-heroku==0.3.1
django-storages==1.13.1
firebase==3.0.1
future==0.16.0
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
jmespath==0.10.0
jwcrypto==1.3
lazy-object-proxy==1.7.1
mccabe==0.7.0
msal==1.18.0
msal-extensions==1.0.0
msgraph-core==0.2.2
oauth2client==4.1.3
oauthlib==3.2.0
pathspec==0.9.0
pbr==5.9.0
perflint==0.7.3
Pillow==9.1.0
platformdirs==2.5.2
portalocker==2.5.1
protobuf==4.21.0rc1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.12.0
PyJWT==2.4.0
pyline==0.3.20
pylint==2.14.5
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.2
pytz==2022.4
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rsa==4.8
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==4.0.0
termcolor==1.1.0
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.1.9
whitenoise==6.2.0
wrapt==1.14.1
